Source/ECGSim_SW/requirements.txt
pyserial
numpy
neurokit2matplotlib